requirements.txt
# Name                    Version
matplotlib                3.5.0
numpy                     1.21.2
pandas                    1.3.5
python                    3.9.7
pytorch                   1.11.0
scipy                     1.7.3
torchaudio                0.11.0
torchvision               0.12.0